requirements.txt
Django
djangorestframework
psycopg2
djangorestframework-simplejwt
python-dotenv